+++ ReciclarteAPI/Models/Materials.cs
@@ -12,7 +12,7 @@ namespace ReciclarteAPI.Models
         public long Id { set; get; }
         [JsonProperty(PropertyName = "name")]
         public string Material { set; get; }
-        public long Price { set; get; }
+        public double Price { set; get; }
         [JsonIgnore]
         public List<MaterialsPerCenter> MaterialsPerCenters { get; set; }
     }
